Move HasJavaClassHashes (and :classhash) up into jvm.core
Summary:
This is jvm rather than java language specific, so move it into core

Test plan: CI
