+++ src/com/facebook/buck/cli/BUCK
@@ -127,8 +127,8 @@ java_immutables_library(
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/io:watchman',
     '//src/com/facebook/buck/js:js',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:autodeps',
-    '//src/com/facebook/buck/jvm/java:classhash',
     '//src/com/facebook/buck/jvm/java:config',
     '//src/com/facebook/buck/jvm/java:packagefinder',
     '//src/com/facebook/buck/jvm/java:rules',
