+++ src/com/facebook/buck/command/BUCK
@@ -14,7 +14,7 @@ java_immutables_library(
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:classhash',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:packagefinder',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
