+++ src/com/facebook/buck/gwt/BUCK
@@ -4,7 +4,7 @@ java_library(
   deps = [
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:classhash',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:steps',
     '//src/com/facebook/buck/jvm/java:support',
