+++ src/com/facebook/buck/jvm/core/HasJavaClassHashes.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.facebook.buck.jvm.java;
+package com.facebook.buck.jvm.core;
 
 import com.facebook.buck.model.HasBuildTarget;
 import com.google.common.collect.ImmutableSortedMap;
