+++ src/com/facebook/buck/jvm/java/intellij/BUCK
@@ -14,8 +14,8 @@ java_immutables_library(
     '//src/com/facebook/buck/cxx:rules',
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:autodeps',
-    '//src/com/facebook/buck/jvm/java:classhash',
     '//src/com/facebook/buck/jvm/java:packagefinder',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
