+++ src/com/facebook/buck/maven/BUCK
@@ -13,7 +13,7 @@ java_library(
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/log:api',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:classhash',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/model:model',
