+++ src/com/facebook/buck/rules/macros/BUCK
@@ -7,7 +7,7 @@ java_library(
   deps=[
     '//src/com/facebook/buck/util/environment:platform',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:classhash',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/parser:rule_pattern',
