+++ src/com/facebook/buck/thrift/BUCK
@@ -8,7 +8,7 @@ java_library(
     '//src/com/facebook/buck/android:packageable',
     '//src/com/facebook/buck/cxx:platform',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:classhash',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:config',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:steps',
