+++ test/com/facebook/buck/cli/BUCK
@@ -135,8 +135,8 @@ java_test(
     '//src/com/facebook/buck/httpserver:httpserver',
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/io:watchman',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java/intellij:intellij',
-    '//src/com/facebook/buck/jvm/java:classhash',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/json:json',
