+++ test/com/facebook/buck/jvm/java/BUCK
@@ -18,7 +18,7 @@ java_library(
     '//src/com/facebook/buck/cxx:platform',
     '//src/com/facebook/buck/cxx:rules',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:classhash',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/jvm/java:steps',
@@ -103,7 +103,7 @@ java_test(
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/jvm/java:autodeps',
-    '//src/com/facebook/buck/jvm/java:classhash',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:config',
     '//src/com/facebook/buck/jvm/java:packagefinder',
     '//src/com/facebook/buck/jvm/java:rules',
