+++ test/com/facebook/buck/maven/BUCK
@@ -12,7 +12,7 @@ java_test(
     '//src/com/facebook/buck/file:downloader',
     '//src/com/facebook/buck/file:file',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:classhash',
+    '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/json:json',
